refactor(backend): extract request reads into query
